refactor: extract reachability func and rework tests
